added licence message

+++ dynamic_range_meter.py
@@ -1,3 +1,19 @@
+# dr14_t.meter: compute the DR14 value of the given audiofiles
+#Copyright (C) 2011  Simone Riva
+#
+#This program is free software: you can redistribute it and/or modify
+#it under the terms of the GNU General Public License as published by
+#the Free Software Foundation, either version 3 of the License, or
+#(at your option) any later version.
+#
+#This program is distributed in the hope that it will be useful,
+#but WITHOUT ANY WARRANTY; without even the implied warranty of
+#MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE.  See the
+#GNU General Public License for more details.
+#
+#You should have received a copy of the GNU General Public License
+#along with this program.  If not, see <http://www.gnu.org/licenses/>.
+
 import read_wav
 import os
 from compute_dr14 import compute_dr14
@@ -37,7 +53,7 @@ class DynamicRangeMeter:
                 res = { 'file_name': file_name , 'dr14': dr14 , 'dB_peak': dB_peak , 'dB_rms': dB_rms }
                 self.res_list.append(res)
              
-        self.dr14 = round( self.dr14 / len( self.res_list ) )
+        self.dr14 = int( round( self.dr14 / len( self.res_list ) ) )
         return len( self.res_list )
